SPARKC-262: Set write consistency level to LOCAL_QUORUM.
With the current default write consistency level of LOCAL_ONE,
it's relatively easy to overwhelm a cluster because it only waits
for one replica to ack the write before proceeding on.
Changing the default write consistency level to LOCAL_QUORUM means
that it will do a little more natural throttling to allow the cluster to keep up
with the writes for multiple replicas.
It seems like it would make for a more stable out of the box experience.