SPARKC-277: Fix decimal Dataframe Test
There were two errors here, first the BigDecimals were being created
with doubles which led to an inevtible loss of precision when
transfering to Dataframes (with fixed precision). This is fixed by
passing in Strings instead; Second, the
java.math.BigDecimal type is not "equal" unless the other BigDecmial has
the same precision as well as value. Switching this to Scala BigDecmial
fixes this behvaior.