SPARKC-277: Adds Nanosecond C* Time Type
We have to be careful with this type because it represents time in
nanoseconds versus Milliseconds like almost all of the other formats.
This means it requires a special converter and we need to be sure that
any transitions from other Date values to Long don't get placed in this
field. For general usage only using Long is the safest bet.