SPARKC-277: Adding a Date -> Time Converter, Warning for Time -> Date
Since our type conversions make it difficult to tell whether a Long was
read from a Time column or a Timestamp column, there will be a factor of
a Million error if a Date value is read out of a Time column. I'm
leaving a warning in the documentation to not perform this conversion.