SPARKC-277: Adjust Timestamps to Match SparkSQL/java.sql.Date Wrapping
java.sql.Date assumes a wrapping of the internal epoch timestamp by the
instantiating JVM. Spark SQL has decided that in there operation they
will take this timestamp and remove the timezone offset before rounding
to determine the Integer representation of the Date. We will have to
perfrom the shift that they expect when going from our UTC based
timestamps coming out of LocalDate to the Locale Timezone. In addition,
this patch introduces converters to do the same thing if an end user is
reading LocalDate objects to java.sql.Date objects.